As a user, I can see many boxes on the page all acting independently of one another.
